3: 3: 3: Create custom hook for fetching & processing data from raider.io
